Add mouseover event listeners
Create a function that leaves a trail (blue background color) when the
mouse enters the grid items.

Optionally, a "mouseout"-event listener was created if no trail is
wanted and only the current grid that the mouse enters should change
its background color.
This hover effect can also be achieved in CSS with a simple statement,
this option uses way less resources and is preferable.